Refactor(#8): 클래스 이름 수정
