--- src/main/kotlin/com/wypl/notification/global/kafka/KafkaConsumerMessage.kt
@@ -1,3 +1,0 @@
-package com.wypl.notification.global.kafka
-
-open class KafkaConsumerMessage() : KafkaMessage
