+++ src/main/kotlin/com/wypl/notification/global/kafka/KafkaEvent.kt
@@ -1,0 +1,3 @@
+package com.wypl.notification.global.kafka
+
+interface KafkaEvent
