+++ src/main/kotlin/com/wypl/notification/global/kafka/KafkaProducerEvent.kt
@@ -1,0 +1,3 @@
+package com.wypl.notification.global.kafka
+
+open class KafkaProducerEvent() : KafkaEvent
