+++ src/main/kotlin/com/wypl/notification/group/event/GroupKafkaConsumerEventType.kt
@@ -1,0 +1,5 @@
+package com.wypl.notification.group.event
+
+enum class GroupKafkaConsumerEventType {
+    GROUP_INVITE,
+}
