--- src/main/kotlin/com/wypl/notification/group/message/GroupKafkaConsumerMessageType.kt
@@ -1,5 +1,0 @@
-package com.wypl.notification.group.message
-
-enum class GroupKafkaConsumerMessageType {
-    GROUP_INVITE,
-}
