+++ src/main/kotlin/com/wypl/notification/schedule/event/ScheduleKafkaConsumerEventType.kt
@@ -1,0 +1,5 @@
+package com.wypl.notification.schedule.event
+
+enum class ScheduleKafkaConsumerEventType {
+    WRITE_REVIEW,
+}
