+++ src/main/kotlin/com/wypl/notification/schedule/event/WriteScheduleReviewEvent.kt
@@ -1,11 +1,11 @@
-package com.wypl.notification.schedule.message
+package com.wypl.notification.schedule.event
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import com.wypl.notification.global.kafka.KafkaConsumerMessage
+import com.wypl.notification.global.kafka.KafkaConsumerEvent
 
-data class WriteScheduleReviewMessage(
+data class WriteScheduleReviewEvent(
     @JsonProperty("recipient_id") val recipientId: Long,
     @JsonProperty("nickname") val nickname: String,
     @JsonProperty("schedule_id") val scheduleId: Long,
     @JsonProperty("schedule_title") val scheduleTitle: String,
-) : KafkaConsumerMessage()
+) : KafkaConsumerEvent()
